chore: componetized functions
